refactor: move configmap client into k8s adapter package
## Summary

Move the Kubernetes configmap client into the main `k8s` adapter package.

## Why

ConfigMaps are Kubernetes-specific outbound adapter concerns, so they belong in the `k8s` package rather than in a separate generic-looking subpackage. This also prepares the package for a shared test suite and a cleaner Testify migration.

Issue: #184
